Update package versions and refactor scope handling
- Bump package versions in project files to 1.4.2.
- Change import statements in `GetScopesRequest.cs` and `FindAccessRoleQuery.cs`.
- Simplify `IRoleRequirement.Roles` property across multiple classes.
- Add `GetPagedScopesQuery` and `GetPagedScopesQueryHandler` for paginated scope retrieval.
- Update `Program.cs` to register new roles for `SystemRoles`.
