+++ CF.Authentication/CF.Identity/CF.Identity.Api/CF.Identity.Api.csproj
@@ -8,9 +8,9 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="IDFCR.Http.Authentication" Version="1.4.1" />
-    <PackageReference Include="IDFCR.Shared.FluentValidation" Version="1.4.1" />
-    <PackageReference Include="IDFCR.Shared.Http" Version="1.4.1" />
+    <PackageReference Include="IDFCR.Http.Authentication" Version="1.4.2" />
+    <PackageReference Include="IDFCR.Shared.FluentValidation" Version="1.4.2" />
+    <PackageReference Include="IDFCR.Shared.Http" Version="1.4.2" />
     <PackageReference Include="FluentValidation.DependencyInjectionExtensions" Version="12.0.0" />
     <PackageReference Include="MediatR" Version="12.5.0" />
     <PackageReference Include="Swashbuckle.AspNetCore" Version="8.1.2" />
