+++ CF.Authentication/CF.Identity/CF.Identity.Api/Endpoints/Scopes/Get/GetScopesRequest.cs
@@ -1,4 +1,4 @@
-﻿using CF.Identity.Api.Features.Scopes.Get;
+﻿using CF.Identity.Api.Features.Scopes.List;
 using CF.Identity.Infrastructure.Features.Scope;
 using IDFCR.Shared.Abstractions.Records;
