+++ CF.Authentication/CF.Identity/CF.Identity.Api/Features/AccessRoles/Find/FindAccessRoleQuery.cs
@@ -1,5 +1,4 @@
-﻿using CF.Identity.Infrastructure.Features;
-using CF.Identity.Infrastructure.Features.AccessRoles;
+﻿using CF.Identity.Infrastructure.Features.AccessRoles;
 using IDFCR.Shared.Abstractions;
 using IDFCR.Shared.Abstractions.Roles;
 using IDFCR.Shared.Mediatr;
@@ -8,6 +7,6 @@ namespace CF.Identity.Api.Features.AccessRoles.Find;
 
 public record FindAccessRoleQuery(Guid AccessRoleId, bool Bypass = false) : IUnitRequest<AccessRoleDetail>, IRoleRequirement
 {
-    IEnumerable<string> IRoleRequirement.Roles => RoleRegistrar.List<Roles>(RoleCategory.Read, SystemRoles.GlobalRead);
+    IEnumerable<string> IRoleRequirement.Roles => RoleRegistrar.List<Roles>(RoleCategory.Read);
     RoleRequirementType IRoleRequirement.RoleRequirementType => RoleRequirementType.Some;
 }
