+++ CF.Authentication/CF.Identity/CF.Identity.Api/Features/AccessRoles/Get/GetAccessRolesQuery.cs
@@ -9,6 +9,6 @@ namespace CF.Identity.Api.Features.AccessRoles.Get;
 public record GetAccessRolesQuery(Guid? ClientId = null, string? Name = null, string? NameContains = null,
     bool NoTracking = true, bool Bypass = false) : IUnitRequestCollection<AccessRoleDto>, IAccessRoleFilter, IRoleRequirement
 {
-    IEnumerable<string> IRoleRequirement.Roles => RoleRegistrar.List<Roles>(RoleCategory.Read, SystemRoles.GlobalRead);
+    IEnumerable<string> IRoleRequirement.Roles => RoleRegistrar.List<Roles>(RoleCategory.Read);
     RoleRequirementType IRoleRequirement.RoleRequirementType => RoleRequirementType.Some;
 }
