+++ CF.Authentication/CF.Identity/CF.Identity.Api/Features/AccessRoles/Post/PostAccessRoleCommand.cs
@@ -7,7 +7,6 @@ namespace CF.Identity.Api.Features.AccessRoles.Post;
 
 public record PostAccessRoleCommand(EditableAccessRoleDto AccessRole, bool Bypass = false) : IUnitRequest<Guid>, IRoleRequirement
 {
-    IEnumerable<string> IRoleRequirement.Roles => RoleRegistrar
-        .List<Roles>(IDFCR.Shared.Abstractions.RoleCategory.Write, SystemRoles.GlobalWrite);
+    IEnumerable<string> IRoleRequirement.Roles => RoleRegistrar.List<Roles>(IDFCR.Shared.Abstractions.RoleCategory.Write);
     RoleRequirementType IRoleRequirement.RoleRequirementType => RoleRequirementType.Some;
 }
