+++ CF.Authentication/CF.Identity/CF.Identity.Api/Features/AccessRoles/Upsert/UpsertAccessRoleCommand.cs
@@ -8,6 +8,6 @@ namespace CF.Identity.Api.Features.AccessRoles.Upsert;
 public record UpsertAccessRoleCommand(EditableAccessRoleDto AccessRole, bool Bypass = false) : IUnitRequest<Guid>, IRoleRequirement
 {
     IEnumerable<string> IRoleRequirement.Roles => RoleRegistrar
-        .List<Roles>(IDFCR.Shared.Abstractions.RoleCategory.Write, SystemRoles.GlobalWrite);
+        .List<Roles>(IDFCR.Shared.Abstractions.RoleCategory.Write);
     RoleRequirementType IRoleRequirement.RoleRequirementType => RoleRequirementType.Some;
 }
