+++ CF.Authentication/CF.Identity/CF.Identity.Api/Features/Scopes/List/GetPagedScopesQueryHandler.cs
@@ -3,7 +3,7 @@ using IDFCR.Shared.Abstractions.Results;
 using IDFCR.Shared.Extensions;
 using IDFCR.Shared.Mediatr;
 
-namespace CF.Identity.Api.Features.Scopes.Get;
+namespace CF.Identity.Api.Features.Scopes.List;
 
 public class GetPagedScopesQueryHandler(IScopeRepository scopeRepository) : IUnitPagedRequestHandler<GetPagedScopesQuery, ScopeDto>
 {
