+++ CF.Authentication/CF.Identity/CF.Identity.Api/Features/Scopes/Upsert/UpsertScopeCommand.cs
@@ -1,6 +1,6 @@
 ﻿using CF.Identity.Api.Features.Scopes.Post;
-using CF.Identity.Infrastructure.Features;
 using CF.Identity.Infrastructure.Features.Scope;
+using IDFCR.Shared.Abstractions;
 using IDFCR.Shared.Abstractions.Results;
 using IDFCR.Shared.Abstractions.Roles;
 using IDFCR.Shared.Mediatr;
@@ -9,7 +9,7 @@ namespace CF.Identity.Api.Features.Scopes.Upsert;
 
 public record UpsertScopeCommand(EditableScopeDto Scope, bool Bypass = false) : IUnitRequest<Guid>, IRoleRequirement
 {
-    IEnumerable<string> IRoleRequirement.Roles => [SystemRoles.GlobalWrite, ScopeRoles.ScopeWrite];
+    IEnumerable<string> IRoleRequirement.Roles => RoleRegistrar.List<ScopeRoles>(RoleCategory.Write);
     RoleRequirementType IRoleRequirement.RoleRequirementType => RoleRequirementType.Some;
 }
