+++ CF.Authentication/CF.Identity/CF.Identity.Api/Features/User/Assign/AssignUserScopesCommand.cs
@@ -1,6 +1,4 @@
-﻿using CF.Identity.Infrastructure.Features;
-using CF.Identity.Infrastructure.Features.Scope;
-using CF.Identity.Infrastructure.Features.Users;
+﻿using CF.Identity.Infrastructure.Features.Users;
 using IDFCR.Shared.Abstractions.Roles;
 using IDFCR.Shared.Abstractions;
 using IDFCR.Shared.Mediatr;
@@ -9,6 +7,6 @@ namespace CF.Identity.Api.Features.User.Assign;
 
 public record AssignUserScopesCommand(Guid ClientId, Guid UserId, IEnumerable<string> Scopes, bool Bypass = false) : IUnitRequest, IRoleRequirement
 {
-    IEnumerable<string> IRoleRequirement.Roles => [ScopeRoles.ScopeRead, UserRoles.UserWrite];
+    IEnumerable<string> IRoleRequirement.Roles => RoleRegistrar.List<UserRoles>(RoleCategory.Write);
     RoleRequirementType IRoleRequirement.RoleRequirementType => RoleRequirementType.Some;
 }
