+++ CF.Authentication/CF.Identity/CF.Identity.Api/Features/User/Get/FindUserByIdQuery.cs
@@ -1,5 +1,5 @@
-﻿using CF.Identity.Infrastructure.Features;
-using CF.Identity.Infrastructure.Features.Users;
+﻿using CF.Identity.Infrastructure.Features.Users;
+using IDFCR.Shared.Abstractions;
 using IDFCR.Shared.Abstractions.Roles;
 using IDFCR.Shared.Mediatr;
 
@@ -7,6 +7,6 @@ namespace CF.Identity.Api.Features.User.Get;
 
 public record FindUserByIdQuery(Guid Id, bool Bypass = false) : IUnitRequest<UserDto>, IRoleRequirement
 {
-    IEnumerable<string> IRoleRequirement.Roles => [SystemRoles.GlobalRead, UserRoles.UserRead];
+    IEnumerable<string> IRoleRequirement.Roles => RoleRegistrar.List<UserRoles>(RoleCategory.Read);
     RoleRequirementType IRoleRequirement.RoleRequirementType => RoleRequirementType.Some;
 }
