+++ CF.Authentication/CF.Identity/CF.Identity.Api/Features/User/Post/PostUserCommand.cs
@@ -1,5 +1,5 @@
-﻿using CF.Identity.Infrastructure.Features;
-using CF.Identity.Infrastructure.Features.Users;
+﻿using CF.Identity.Infrastructure.Features.Users;
+using IDFCR.Shared.Abstractions;
 using IDFCR.Shared.Abstractions.Roles;
 using IDFCR.Shared.Mediatr;
 
@@ -8,6 +8,6 @@ namespace CF.Identity.Api.Features.User.Post;
 public record PostUserCommand(EditableUserDto User) : IUnitRequest<Guid>, IRoleRequirement
 {
     bool IRoleRequirement.Bypass => false;
-    IEnumerable<string> IRoleRequirement.Roles => [SystemRoles.GlobalWrite, UserRoles.UserWrite];
+    IEnumerable<string> IRoleRequirement.Roles => RoleRegistrar.List<UserRoles>(RoleCategory.Write);
     RoleRequirementType IRoleRequirement.RoleRequirementType => RoleRequirementType.Some;
 }
