+++ CF.Authentication/CF.Identity/CF.Identity.Infrastructure.SqlServer/CF.Identity.Infrastructure.SqlServer.csproj
@@ -6,7 +6,7 @@
     <Nullable>enable</Nullable>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="IDFCR.Shared.EntityFramework" Version="1.4.1" />
+    <PackageReference Include="IDFCR.Shared.EntityFramework" Version="1.4.2" />
     <PackageReference Include="LinqKit.Core" Version="1.2.8" />
     <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="9.0.5" />
     <PackageReference Include="Scrutor" Version="6.0.1" />
