+++ CF.Authentication/CF.Identity/CF.Identity.Infrastructure/CF.Identity.Infrastructure.csproj
@@ -6,7 +6,7 @@
     <Nullable>enable</Nullable>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="IDFCR.Shared" Version="1.4.1" />
+    <PackageReference Include="IDFCR.Shared" Version="1.4.2" />
     <PackageReference Include="System.IdentityModel.Tokens.Jwt" Version="8.11.0" />
     <PackageReference Include="Microsoft.Extensions.Configuration.Binder" Version="9.0.5" />
   </ItemGroup>
