+++ CF.Authentication/CF.Identity/CF.Identity.MigrationUtility/CF.Identity.MigrationUtility.csproj
@@ -9,7 +9,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="IDFCR.Utility.Shared" Version="1.4.1" />
+    <PackageReference Include="IDFCR.Utility.Shared" Version="1.4.2" />
     <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="9.0.5">
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers</IncludeAssets>
